modules/data-ingestion/requirements.txt
earthengine-api==0.1.384
geopandas==0.14.1
rasterio==1.3.9
numpy==1.26.2
pandas==2.1.3
xarray==2023.11.0
netCDF4==1.6.5
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
